feat: 소셜 콜백 스켈레톤/스피너 추가 및 홈 가드, 404, 메타데이터, 로그인 설정 개선

+++ hooks/useCalcPace.ts
@@ -8,10 +8,10 @@ export function useCalcPace(keepPaceOnEmpty: boolean = false) {
   const [pace, setPace] = useState("");
 
   useEffect(() => {
-    const h = parseInt(hour) || 0;
-    const m = parseInt(min) || 0;
-    const s = parseInt(sec) || 0;
-    const d = parseInt(distance) || 0;
+    const h = parseFloat(hour) || 0;
+    const m = parseFloat(min) || 0;
+    const s = parseFloat(sec) || 0;
+    const d = parseFloat(distance) || 0;
     if (d > 0 && (h > 0 || m > 0 || s > 0)) {
       const totalMinutes = h * 60 + m + s / 60;
       const paceInMinutes = totalMinutes / d;
